requirements.txt
discord.py==2.5.2
python-dotenv==1.1.0
yt-dlp==2025.3.31
lyricsgenius==3.6.2
